refactor: chart auto install series
