+++ packages/vchart/src/chart/bar/bar.ts
@@ -2,6 +2,9 @@ import { SeriesTypeEnum } from '../../series/interface';
 import { CartesianChart } from '../cartesian/cartesian';
 import { ChartTypeEnum } from '../interface';
 import { setDefaultCrosshairForCartesianChart } from '../util';
+import { VChart } from '../../core/vchart';
+import { BarSeries } from '../../series';
+VChart.useSeries([BarSeries]);
 
 export class BarChart extends CartesianChart {
   static readonly type: string = ChartTypeEnum.bar;
@@ -13,11 +16,4 @@ export class BarChart extends CartesianChart {
     super.transformSpec(spec);
     setDefaultCrosshairForCartesianChart(spec);
   }
-}
-
-export class Bar3dChart extends CartesianChart {
-  static readonly type: string = ChartTypeEnum.bar3d;
-  static readonly view: string = 'singleDefault';
-  readonly type: string = ChartTypeEnum.bar3d;
-  readonly seriesType: string = SeriesTypeEnum.bar3d;
 }
