+++ packages/vchart/src/chart/bar/index.ts
@@ -1,2 +1,3 @@
 export * from './bar';
+export * from './bar-3d';
 export * from './interface';
