+++ packages/vchart/src/chart/funnel/index.ts
@@ -1,2 +1,3 @@
 export * from './funnel';
+export * from './funnel-3d';
 export * from './interface';
