+++ packages/vchart/src/chart/heatmap/heatmap.ts
@@ -1,6 +1,9 @@
 import { SeriesTypeEnum } from '../../series/interface';
 import { CartesianChart } from '../cartesian/cartesian';
 import { ChartTypeEnum } from '../interface';
+import { VChart } from '../../core/vchart';
+import { HeatmapSeries } from '../../series';
+VChart.useSeries([HeatmapSeries]);
 
 export class HeatmapChart extends CartesianChart {
   static readonly type: string = ChartTypeEnum.heatmap;
