+++ packages/vchart/src/chart/histogram/histogram.ts
@@ -1,23 +1,10 @@
-import type { ICartesianAxisCommonSpec } from '../../component/axis/cartesian/interface';
 import { SeriesTypeEnum } from '../../series/interface';
-import { CartesianChart } from '../cartesian/cartesian';
 import { ChartTypeEnum } from '../interface';
+import { BaseHistogramChart } from './base';
 import { setDefaultCrosshairForCartesianChart } from '../util';
-
-class BaseHistogramChart extends CartesianChart {
-  transformSpec(spec: any): void {
-    super.transformSpec(spec);
-    spec.axes.forEach((axis: ICartesianAxisCommonSpec) => (axis.type = 'linear'));
-  }
-
-  protected _getDefaultSeriesSpec(spec: any): any {
-    return {
-      ...super._getDefaultSeriesSpec(spec),
-      x2Field: spec?.x2Field,
-      y2Field: spec?.y2Field
-    };
-  }
-}
+import { VChart } from '../../core/vchart';
+import { BarSeries } from '../../series';
+VChart.useSeries([BarSeries]);
 
 export class HistogramChart extends BaseHistogramChart {
   static readonly type: string = ChartTypeEnum.histogram;
@@ -29,11 +16,4 @@ export class HistogramChart extends BaseHistogramChart {
     super.transformSpec(spec);
     setDefaultCrosshairForCartesianChart(spec);
   }
-}
-
-export class Histogram3dChart extends BaseHistogramChart {
-  static readonly type: string = ChartTypeEnum.histogram3d;
-  static readonly view: string = 'singleDefault';
-  readonly type: string = ChartTypeEnum.histogram3d;
-  readonly seriesType: string = SeriesTypeEnum.bar3d;
 }
