+++ packages/vchart/src/chart/histogram/index.ts
@@ -1,2 +1,3 @@
 export * from './histogram';
+export * from './histogram-3d';
 export * from './interface';
