+++ packages/vchart/src/chart/pie/index.ts
@@ -1,2 +1,3 @@
 export * from './pie';
+export * from './pie-3d';
 export * from './interface';
