+++ packages/vchart/src/chart/range-column/index.ts
@@ -1,2 +1,3 @@
-export * from './rangeColumn';
+export * from './range-column';
+export * from './range-column-3d';
 export * from './interface';
