+++ packages/vchart/src/chart/range-column/range-column.ts
@@ -3,6 +3,9 @@ import { ChartTypeEnum } from '../interface';
 import { SeriesTypeEnum } from '../../series/interface';
 import { Direction } from '../../typings';
 import { setDefaultCrosshairForCartesianChart } from '../util';
+import { VChart } from '../../core/vchart';
+import { RangeColumnSeries } from '../../series';
+VChart.useSeries([RangeColumnSeries]);
 
 export class RangeColumnChart extends CartesianChart {
   static readonly type: string = ChartTypeEnum.rangeColumn;
@@ -26,25 +29,5 @@ export class RangeColumnChart extends CartesianChart {
   transformSpec(spec: any): void {
     super.transformSpec(spec);
     setDefaultCrosshairForCartesianChart(spec);
-  }
-}
-
-export class RangeColumn3dChart extends CartesianChart {
-  static readonly type: string = ChartTypeEnum.rangeColumn3d;
-  static readonly view: string = 'singleDefault';
-  readonly type: string = ChartTypeEnum.rangeColumn3d;
-  readonly seriesType: string = SeriesTypeEnum.rangeColumn3d;
-
-  protected _getDefaultSeriesSpec(spec: any): any {
-    const series: any = {
-      ...super._getDefaultSeriesSpec(spec)
-    };
-    series.bar3d = spec.bar3d;
-    if (spec.direction === Direction.horizontal) {
-      series.xField = spec.xField ?? [spec.minField, spec.maxField];
-    } else {
-      series.yField = spec.yField ?? [spec.minField, spec.maxField];
-    }
-    return series;
   }
 }
