+++ packages/vchart/src/chart/scatter/scatter.ts
@@ -2,6 +2,9 @@ import { SeriesTypeEnum } from '../../series/interface';
 import { CartesianChart } from '../cartesian/cartesian';
 import { ChartTypeEnum } from '../interface';
 import type { IScatterChartSpec } from './interface';
+import { VChart } from '../../core/vchart';
+import { ScatterSeries } from '../../series';
+VChart.useSeries([ScatterSeries]);
 
 export class ScatterChart extends CartesianChart {
   static readonly type: string = ChartTypeEnum.scatter;
