+++ packages/vchart/src/chart/waterfall/waterfall.ts
@@ -4,6 +4,9 @@ import { BarChart } from '../bar';
 import { ChartTypeEnum } from '../interface';
 import type { IWaterfallChartSpec } from './interface';
 import { setDefaultCrosshairForCartesianChart } from '../util';
+import { VChart } from '../../core/vchart';
+import { WaterfallSeries } from '../../series';
+VChart.useSeries([WaterfallSeries]);
 
 export class WaterfallChart extends BarChart {
   static readonly type: string = ChartTypeEnum.waterfall;
