+++ packages/vchart/src/chart/word-cloud/base.ts
@@ -1,0 +1,31 @@
+import { SeriesTypeEnum } from '../../series/interface';
+import { ChartTypeEnum } from '../interface';
+// eslint-disable-next-line no-duplicate-imports
+import type { ISeries } from '../../series/interface';
+import { BaseChart } from '../base-chart';
+
+export class BaseWordCloudChart extends BaseChart {
+  readonly type: string = ChartTypeEnum.wordCloud;
+  readonly seriesType: string = SeriesTypeEnum.wordCloud;
+
+  transformSpec(spec: any): void {
+    super.transformSpec(spec);
+
+    /* 处理 series 配置 */
+    const defaultSeriesSpec = this._getDefaultSeriesSpec(spec);
+    if (!spec.series || spec.series.length === 0) {
+      spec.series = [defaultSeriesSpec];
+    } else {
+      spec.series.forEach((s: ISeries) => {
+        if (!this.isValidSeries(s.type)) {
+          return;
+        }
+        Object.keys(defaultSeriesSpec).forEach(k => {
+          if (!(k in s)) {
+            s[k] = defaultSeriesSpec[k];
+          }
+        });
+      });
+    }
+  }
+}
