+++ packages/vchart/src/chart/word-cloud/index.ts
@@ -1,2 +1,3 @@
 export * from './word-cloud';
+export * from './word-cloud-3d';
 export * from './interface';
